Update nextcloud-18-installation-on-ubuntu.md
Nextcloud update to 20.0.5 flags this error with the install method above:

"Security & setup warnings

It's important for the security and performance of your instance that everything is configured correctly. To help you with that we are doing some automatic checks. Please see the linked documentation for more information.
There are some warnings regarding your setup.

    The "X-Robots-Tag" HTTP header is not set to "noindex, nofollow". This is a potential security or privacy risk, as it is recommended to adjust this setting accordingly."

Updating the sites-available as above gets rid of this error message.
You may want to use this change?